+++ lib/BeGateway/GatewayTransport.php
@@ -3,7 +3,7 @@ namespace BeGateway;
 
 class GatewayTransport {
 
-    public static function submit($shop_id, $shop_key, $host, $t_request, $curl_timeout = 30, $curl_connect_timeout = 10) {
+    public static function submit($shop_id, $shop_key, $host, $t_request, $headers = array(), $curl_timeout = 30, $curl_connect_timeout = 10) {
 
         $process = curl_init($host);
         $json = json_encode($t_request);
@@ -14,11 +14,21 @@ class GatewayTransport {
           Logger::getInstance()->write("with message " .  $json, Logger::DEBUG, get_class());
 
         if (!empty($t_request)) {
-          curl_setopt($process, CURLOPT_HTTPHEADER, array('Accept: application/json', 'Content-type: application/json'));
+          curl_setopt($process, CURLOPT_HTTPHEADER,
+            array_merge(
+              array('Accept: application/json', 'Content-type: application/json'),
+              $headers
+            )
+          );
           curl_setopt($process, CURLOPT_POST, 1);
           curl_setopt($process, CURLOPT_POSTFIELDS, $json);
         } else {
-          curl_setopt($process, CURLOPT_HTTPHEADER, array('Accept: application/json'));
+          curl_setopt($process, CURLOPT_HTTPHEADER,
+            array_merge(
+              array('Accept: application/json'),
+              $headers
+            )
+          );
         }
 
         curl_setopt($process, CURLOPT_URL, $host);